Fully fixed deleting/adding/updating Cryptos and Portfolio Entries, Value Checks currently bugged

+++ PortfolioApp/portfolio_tracker.py
@@ -12,7 +12,6 @@ from kivy.core.window import Window
 from kivy.uix.screenmanager import Screen, ScreenManager
 from kivy_garden.matplotlib import FigureCanvasKivyAgg
 from matplotlib import pyplot as plt
-from sqlalchemy import desc, ForeignKeyConstraint
 from sqlalchemy.exc import SQLAlchemyError
 
 from HistoricalPriceViewer.main import coin_gecko_api
@@ -99,14 +98,12 @@ class AddCryptocurrencyScreen(Screen):
 class DeleteCryptocurrencyScreen(Screen):
     def reset_page(self):
         self.ids.spinner_delete_crypto.text = 'Crypto ID' if len(self.ids.spinner_delete_crypto.text) > 0 else 'No Cryptos'
-        self.ids.delete_crypto_name.text = 'Crypto Name'
-        self.ids.delete_crypto_symbol.text = 'Crypto Symbol'
+        self.ids.delete_crypto_name.text = ''
+        self.ids.delete_crypto_symbol.text = ''
 
     def update_page(self):
-        print('updating DeleteCrypto screen')
         spinner_text = self.ids.spinner_delete_crypto.text
         if spinner_text == 'Crypto ID' or spinner_text == 'No Cryptos':
-            print('not updating')
             return
         crypto = app.session.query(Crypto).filter(Crypto.crypto_id == spinner_text).first()
         self.ids.delete_crypto_name.text = crypto.name
@@ -139,14 +136,12 @@ class UpdatePortfolioEntryScreen(Screen):
     def update_page(self):
         # Retrieve the entry with the same id as the user selects from the spinner
         spinner_text = self.ids.spinner_update_portfolio.text
-        numbers_in_spinner_text = re.match('\d', spinner_text)
+        numbers_in_spinner_text = re.match(r'\d', spinner_text)
         if numbers_in_spinner_text is None:
             return
-        print(spinner_text, numbers_in_spinner_text)
         entry = app.session.query(PortfolioEntry).filter(PortfolioEntry.entry_id == int(spinner_text)).first()
         self.ids.update_spinner_crypto_id.text = entry.crypto_id
         self.ids.update_date_text.hint_text = 'YYYY-MM-DD'
-        print(self.ids.update_date_text.text, entry.timestamp.date())
         self.ids.update_date_text.text = str(entry.timestamp.date())
         self.ids.update_quantity_text.hint_text = '1'
         self.ids.update_quantity_text.text = str(entry.quantity)
@@ -162,12 +157,11 @@ class DeletePortfolioEntryScreen(Screen):
     def update_page(self):
         # Retrieve the entry with the same id as the user selects from the spinner
         spinner_text = self.ids.spinner_delete_portfolio.text
-        numbers_in_spinner_text = re.match('\d', spinner_text)
+        numbers_in_spinner_text = re.match(r'\d', spinner_text)
         if numbers_in_spinner_text is None:
             return
         entry = app.session.query(PortfolioEntry).filter(PortfolioEntry.entry_id == int(spinner_text)).first()
         self.ids.delete_crypto_id.text = entry.crypto_id
-        print(self.ids.delete_date_text.text, entry.timestamp.date())
         self.ids.delete_date_text.text = str(entry.timestamp.date())
         self.ids.delete_quantity_text.text = str(entry.quantity)
 
@@ -190,10 +184,6 @@ class PortfolioTrackerApp(App):
     user_id = NumericProperty(1)
 
     background_color = ColorProperty([0.0, 0.0, 0.0, 1.0])
-    text_color = ColorProperty([0.0, 0.0, 0.0, 1.0])
-    outline_color = ColorProperty([1.0, 1.0, 1.0, 1.0])
-    # Button color has to be transparent so its canvas color isn't obscured by it
-    button_color = ColorProperty([1.0, 1.0, 1.0, 0.0])
 
     portfolio_report_date = StringProperty('')
     portfolio_report_previous_date = StringProperty('')
@@ -359,16 +349,27 @@ class PortfolioTrackerApp(App):
         self.display_popup('Entry Added', 'Crypto entry added.', 'Portfolio Menu')
 
     def delete_crypto(self, crypto_id):
+        # Don't delete if the spinner text is not set to a value
+        if crypto_id == 'Crypto ID' or crypto_id == 'No Cryptos':
+            self.display_popup('Empty Data',
+                               'Please select a crypto to delete, if there are any.',
+                               'Delete Cryptocurrency')
+            return
         deleted_crypto = self.session.query(Crypto).filter(Crypto.crypto_id == crypto_id).first()
-        try:
+        affected_entries = self.session.query(PortfolioEntry).filter(PortfolioEntry.crypto_id == crypto_id)
+        # Don't delete a crypto that is used in an entry
+        if affected_entries.count() == 0:
             self.session.delete(deleted_crypto)
             self.session.commit()
-        except SQLAlchemyError as error:
-            print(error)
+        else:
             self.display_popup("Can't Remove Crypto",
                                'This crypto is currently used in a portfolio entry, '
                                'so it cannot be deleted.',
                                next_screen='Delete Cryptocurrency')
+            return
+        self.display_popup('Deleted Crypto',
+                           f'Crypto "{crypto_id}" successfully deleted',
+                           'Portfolio Menu')
 
     def add_crypto_price(self, crypto_id, timestamp, price):
         """
@@ -380,7 +381,6 @@ class PortfolioTrackerApp(App):
         """
         crypto_price = CryptoPrice(crypto_id=crypto_id, timestamp=timestamp, price=price * 100)
         # If an entry in crypto_prices with that timestamp and crypto already exists, exit
-        print(344)
         if self.session.query(CryptoPrice).filter(CryptoPrice.crypto_id == crypto_id).filter(
                 CryptoPrice.timestamp == timestamp).count() != 0:
             return
@@ -405,7 +405,7 @@ class PortfolioTrackerApp(App):
                 entry = self.session.query(PortfolioEntry).filter(PortfolioEntry.entry_id == entry_id).one()
             except SQLAlchemyError:
                 self.display_popup('No Portfolio Entry Found',
-                                   f'There is no portfolio entry with id {entry_id}',
+                                   f'There is no portfolio entry with id "{entry_id}"',
                                    'Manage Portfolio Entries')
                 return
             self.session.delete(entry)
@@ -449,7 +449,6 @@ class PortfolioTrackerApp(App):
             # Add a price entry for the given date if none are found in the Database
             if self.session.query(CryptoPrice).filter(CryptoPrice.timestamp == entry_date).filter(
                     CryptoPrice.crypto_id == crypto_id).count() == 0:
-                print('Line 415, no price entries for this crypto, should add some')
                 try:
                     historic_coin_data = coin_gecko_api.get_coin_history_by_id(crypto_id, adjusted_date)
                     price = historic_coin_data['market_data']['current_price']['usd']
@@ -463,12 +462,7 @@ class PortfolioTrackerApp(App):
                                        'This crypto did not exist at your selected date. Please select a later date.',
                                        f'{operation} Portfolio Entry')
                     return
-                print(self.session.query(CryptoPrice).filter(CryptoPrice.timestamp == entry_date).filter(
-                    CryptoPrice.crypto_id == crypto_id))
             try:
-                prices = self.session.query(CryptoPrice).filter(CryptoPrice.crypto_id == crypto_id)
-                for price in prices:
-                    print(price.timestamp, entry_date, price.timestamp == entry_date)
                 investment = self.session.query(CryptoPrice).filter(CryptoPrice.crypto_id == crypto_id).filter(
                     CryptoPrice.timestamp == entry_date).first().price * int(
                     quantity)
@@ -504,7 +498,7 @@ class PortfolioTrackerApp(App):
                 entry = self.session.query(PortfolioEntry).filter(PortfolioEntry.entry_id == entry_id).one()
             except SQLAlchemyError:
                 self.display_popup('No Portfolio Entry Found',
-                                   f'There is no portfolio entry with id {entry_id}',
+                                   f'There is no portfolio entry with id "{entry_id}"',
                                    'Manage Portfolio Entries')
                 return
             entry.crypto_id = crypto_id
@@ -530,16 +524,13 @@ class PortfolioTrackerApp(App):
         total_value = 0
         total_initial_investment = 0
         crypto_quantities = self.get_quantities_and_investments(timestamp)
-        print(crypto_quantities, timestamp)
         for crypto_id in crypto_quantities:
-            print(496)
             total_initial_investment += crypto_quantities[crypto_id][1]
             total_value += crypto_quantities[crypto_id][2]
 
         change_from_previous = None
 
         if not is_first_value_check:
-            print(503)
             previous_value = previous_value_check.total_value
             change_from_previous = 100 * (total_value - previous_value) // abs(
                 previous_value) if previous_value != 0 else 0
@@ -550,7 +541,6 @@ class PortfolioTrackerApp(App):
                                  change_from_previous=change_from_previous,
                                  user_id=self.user_id, change_from_investment=change_from_investment)
         try:
-            print(514)
             self.session.add(value_check)
             self.session.commit()
         except sqlalchemy.exc.DataError as data_error:
@@ -565,7 +555,6 @@ class PortfolioTrackerApp(App):
                                      user_id=self.user_id, change_from_investment=change_from_investment)
             self.session.add(value_check)
             self.session.commit()
-        print(529)
         self.portfolio_report_date = str(timestamp.date())
         self.portfolio_report_total = total_value
         self.portfolio_report_previous_date = str(
@@ -594,14 +583,10 @@ class PortfolioTrackerApp(App):
             crypto_price = self.session.query(CryptoPrice).filter(CryptoPrice.timestamp == timestamp).filter(
                 CryptoPrice.crypto_id == crypto_id)
             # Add new price to database if none are found at the current time
-            print(558)
             if crypto_price.count() == 0:
-                print(560)
                 price = 100 * round(coin_gecko_api.get_price(crypto_id, 'usd')[crypto_id]['usd'], 2)
                 self.add_crypto_price(crypto_id, timestamp, price)
                 self.session.commit()
-                print(564)
-            print(565)
             current_price = self.session.query(CryptoPrice).filter(CryptoPrice.crypto_id == crypto_id).filter(
                 CryptoPrice.timestamp == timestamp).first()
             crypto_quantities_prices[crypto_id][2] = crypto_quantities_prices[crypto_id][0] * current_price.price
@@ -633,9 +618,5 @@ class CustomButton(Button):
 if __name__ == '__main__':
     Window.size = (400, (16 / 9) * 400)
     Window.top = 0
-    try:
-        app = PortfolioTrackerApp()
-        app.run()
-    except SQLAlchemyError as e:
-        app.session.rollback()
-        print(f'There was a database error: {e}')
+    app = PortfolioTrackerApp()
+    app.run()
